completing middleware set up for authorising permissions

+++ server/routes/projects.ts
@@ -9,16 +9,14 @@ import {
 
 const router = Router()
 
-import { auth, requiredScopes } from 'express-oauth2-jwt-bearer'
+import { auth } from 'express-oauth2-jwt-bearer'
+import { checkPermissions } from 'server/middleware/checkPermissions'
 
 const checkJwt = auth({
   audience: 'https://liamsimpsonportfolio/api',
   issuerBaseURL: 'https://dev-wboo3txpyqmiudzh.au.auth0.com/',
   tokenSigningAlg: 'RS256',
 })
-const checkEditScope = requiredScopes('edit:project')
-const checkAddScope = requiredScopes('add:project')
-const checkDeleteScope = requiredScopes('delete:project')
 
 router.get('/', async (_req, res, next) => {
   try {
@@ -40,35 +38,50 @@ router.get('/:id', async (req, res, next) => {
   }
 })
 
-router.post('/', checkJwt, checkAddScope, async (req, res, next) => {
-  const project = req.body
-  try {
-    await addProject(project)
-    res.sendStatus(201)
-  } catch (error) {
-    next(error)
-  }
-})
+router.post(
+  '/',
+  checkJwt,
+  checkPermissions('add:project'),
+  async (req, res, next) => {
+    const project = req.body
+    try {
+      await addProject(project)
+      res.sendStatus(201)
+    } catch (error) {
+      next(error)
+    }
+  },
+)
 
-router.delete('/:id', checkJwt, checkDeleteScope, async (req, res, next) => {
-  const id = Number(req.params.id)
-  try {
-    await deleteProject(id)
-    res.sendStatus(204)
-  } catch (error) {
-    next(error)
-  }
-})
+router.delete(
+  '/:id',
+  checkJwt,
+  checkPermissions('delete:project'),
+  async (req, res, next) => {
+    const id = Number(req.params.id)
+    try {
+      await deleteProject(id)
+      res.sendStatus(204)
+    } catch (error) {
+      next(error)
+    }
+  },
+)
 
-router.patch('/:id', checkJwt, checkEditScope, async (req, res, next) => {
-  const id = Number(req.params.id)
-  const changes = req.body
-  try {
-    await editProject(id, changes)
-    res.sendStatus(201)
-  } catch (error) {
-    next(error)
-  }
-})
+router.patch(
+  '/:id',
+  checkJwt,
+  checkPermissions('edit:project'),
+  async (req, res, next) => {
+    const id = Number(req.params.id)
+    const changes = req.body
+    try {
+      await editProject(id, changes)
+      res.sendStatus(201)
+    } catch (error) {
+      next(error)
+    }
+  },
+)
 
 export default router
